Refactor: move property providers into providers/properties subdirectory
